examples/xApp/python3/xapp_ricmon/requirements.txt
python-snappy~=0.6.1
requests~=2.22.0
numpy~=1.23.5
polars~=0.17.9